==> diabetesprogression/scripts/04_03_feat_imp_conf_mats.ipynb <==
# Este cuaderno busca el mejor modelo para cada horizonte temporal, estadío base y variable de salida

## Librerias necesarias
import pycaret
from pycaret.classification import *
from pycaret.classification import ClassificationExperiment

import os
import re
import socket
import shutil
import warnings
import numpy as np
import pandas as pd
import plotly.io as pio
import plotly.express as px 
from sklearn.metrics import confusion_matrix

## Configuración de rutas
if socket.gethostname()=='SRVCBECO01':
    os.chdir(r'C:\Users\juanpablo.martinez\Downloads\ryr_exes')
elif socket.gethostname() == 'HPJP2': 
    os.chdir(r'C:\Users\juanm\Dropbox\JP_files\UR\Vías clinicas diabetes')
elif socket.gethostname() == 'CNF106054': 
    os.chdir(r'C:\Users\paul.rodriguez\Dropbox\Salud Colombia\Diabetes Sanitas\Vías clinicas diabetes')
elif socket.gethostname() == 'CNF77701': 
    os.chdir(r'C:\Users\juanpablo.martinez\Dropbox\Vías clinicas diabetes')

    
raw_path = ''
created_path = ""
temp_path = ''
dbs_path = ""


print('Directorio actual: '+ os.getcwd())

### Server packages versions

! pip list

## Funciones

### Definición de parámetros básicos

def parameters_gather(tgt_year, base_label, dep_var):
    
    ## Construyendo los paths de entrenamiento y validación
    train_path = dbs_path+"train\\train_db_{}_years_base_{}.csv".format(tgt_year, base_label)
    vali_path = dbs_path+"vali\\vali_db_{}_years_base_{}.csv".format(tgt_year, base_label)

    ## Listado de variables a eliminar 
    vars_to_drop = ['KeyAnonimo', 'year', 'base_label', 'tgt_label']
    if dep_var == "comorbilidades":
        vars_to_drop = vars_to_drop+['fuera_metas']
    elif dep_var == "fuera_metas":
        vars_to_drop = vars_to_drop+['comorbilidades']
    
    return train_path, vali_path, vars_to_drop

### Preparación de las bases de entrenamiento y validación

def data_prepare(df_path, vars_to_drop):
    ## Cargando la base de datos
    data_mat = pd.read_csv(df_path)
    
    ## Eliminando variables innecesarias
    data_mat.drop(vars_to_drop, axis = 1, inplace = True)
    
    return(data_mat.reset_index(drop = True))

### Identifcando variables categóricas y numéricas en la base de datos

def cat_num_identify(ml_data, rename_df):
    categorical_vars_master = ['femenino', 'ERC_high', 'ERC1', 'ERC2', 'adhiere_guia', 'no_adhiere', 'adhiere_colesterol', 'analgesicos',
                                   'antiacidos', 'antihipertensivos', 'hipoglicemiantes', 'hipolipemiantes', 'nutrition_tag_max', 
                                   'exercise_tag_max', 'alcohol_tag_max', 'tobaco_tag_max', 'creatinina_missing']

    categorical_vars = list(set(ml_data.columns).intersection(categorical_vars_master))

    numerical_vars = set(ml_data.columns).difference(set(categorical_vars+['KeyAnonimo', 'year', 'base_label', 'comorbilidades', 'fuera_metas', 'tgt_label']))
    numerical_vars = list(numerical_vars)

    types_df = pd.DataFrame({'old_name' : categorical_vars+numerical_vars,
                             'type' : ['categorical']*len(categorical_vars)+['numerical']*len(numerical_vars)})

    types_df = types_df.merge(rename_df, on = 'old_name', how = 'inner')

    new_cat_vars = list(types_df.loc[types_df['type'] == 'categorical', 'new_name'].values)
    new_num_vars = list(types_df.loc[types_df['type'] == 'numerical', 'new_name'].values)
    
    return new_cat_vars, new_num_vars

## Parámetros básicos

## Diccionario para renombrar variables
rename_dict ={'femenino' : 'Female', 'edad' : 'Age (Years)', 'peso': 'Weight (Kg)', 'talla' : 'Height (m)', 
              'imc' : "BMI (kg/m*m)", 'Colesterol_LDL' : 'LDL Chol. (mg/dL)', 'TFG' : "eGFR (mg/g)", 
              "ta_diastolica" : "Diast. B.P. (mmHg)", 'ta_sistolica' : 'Sist. B.P. (mmHg)', 'adhiere_guia' : "Hba1c guide Adh.", 
              "no_adhiere" : "Pharma. Adh.", "analgesicos" : "Analgesics", "antiacidos" : "Antacids",
              "antihipertensivos" : "Antihypertensive", "hipoglicemiantes" : "Hypoglecimic agents", 
              "hipolipemiantes" : "Lipid-lowering agents", "nutrition_tag_max" : "Nutrition recomm.", 
              "exercise_tag_max" : "Physical act. recomm.", "alcohol_tag_max" : "Alcohol recomm.",
              "tobaco_tag_max" : "Tobacco recomm.", "creatinina" : "Creatinine (mg/dL)", 
              "adhiere_colesterol" : "Chol. Adh.", 
             "ERC1" : "CKD-1", "ERC2" : "CKD-2", "ERC3" : "CKD-Late"} 

#'imc': r"BMI ($\displaystyle\frac{kg}{m^2}$)"}

rename_df = pd.DataFrame(rename_dict, index = rename_dict.keys())
rename_df = pd.DataFrame({'old_name' : rename_df.index, 'new_name' : np.diag(rename_df)})

## Comparando los modelos

### Cargando la info. de Pycaret

## Load all the .csv files
all_results = []
for file_name in [each for each in os.listdir(temp_path+"ryr_results/baseline_replicate/") if each.endswith('.csv')]:
    all_results.append(pd.read_csv(temp_path+"ryr_results/baseline_replicate/"+file_name, sep = ';'))

## Append all the results
all_results = pd.concat(all_results).reset_index(drop = True) 

appendix_table = all_results[all_results['samp_type'] != "train_manual"]

# Isolating the best models in terms of F1 
cols_order = ['tgt_year', 'base_label', 'dep_var', 'samp_type']+['Model', "Accuracy", "AUC", "F1", "Kappa"]
best_models_index = all_results.loc[all_results['samp_type'] == "training", 
                                    ['tgt_year', 
                                     'base_label', 
                                     'dep_var', 
                                     'F1']].groupby(['tgt_year', 
                                                     'base_label', 
                                                     'dep_var']).idxmax().reset_index(drop = True)["F1"]

## Keep the validation result and the best training result
ordered_results = pd.concat([all_results.loc[best_models_index, cols_order], 
           all_results.loc[all_results['samp_type'] == "validation", cols_order]])

### Cargando la info. de las  redes neuronales y otros modelos específico

## Gathering the info. from the Neural Networks
dep_var = "fuera_metas"

# Iterate across years and initial statess
for tgt_year in [1, 2]:
    for base_label in [1, 3]:
        # Load the Neural Networks results
        nn_models = pd.read_csv(temp_path+'nn_res_{}_{}_years_base_{}.csv'.format(dep_var, tgt_year, base_label))
        nn_models['Model'] = "NN"

        # Iterate across possible samples
        for tgt_samp in ["training", "validation"]:
            
            best_model = nn_models.loc[(nn_models['samp_type'] == tgt_samp), 'f1'].idxmax()

            # Manually replace the Model type, accuracy, AUC and F1 score
            index_to_replace = (ordered_results['dep_var'] == dep_var) & (ordered_results['tgt_year'] == tgt_year) & (ordered_results['base_label'] == base_label) & (ordered_results['samp_type'] == tgt_samp)
            ordered_results.loc[index_to_replace, "Model"] = "NN"
            ordered_results.loc[index_to_replace, "Accuracy"] = nn_models.loc[best_model, "accuracy"]
            ordered_results.loc[index_to_replace, "AUC"] = nn_models.loc[best_model, "auc"]
            ordered_results.loc[index_to_replace, "F1"] = nn_models.loc[best_model, "f1"]
            
dep_var = "fuera_metas"
tgt_year = 2
base_label = 2
tgt_model = "Light Gradient Boosting Machine"
for tgt_samp in ["training", "validation"]:

    # Manually replace the Model type, accuracy, AUC and F1 score
    best_model = pd.read_csv(temp_path+"org_{}_years_base_{}_{}.csv".format(tgt_year, base_label, dep_var), sep = ';')
    best_model = best_model[(best_model['samp_type'] == tgt_samp) & (best_model['Model'] == tgt_model)].reset_index(drop = True)

    index_to_replace = (ordered_results['dep_var'] == dep_var) & (ordered_results['tgt_year'] == tgt_year) & (ordered_results['base_label'] == base_label) & (ordered_results['samp_type'] == tgt_samp)
    ordered_results.loc[index_to_replace, "Model"] = best_model.loc[0, "Model"]
    ordered_results.loc[index_to_replace, "Accuracy"] = best_model.loc[0, "Accuracy"]
    ordered_results.loc[index_to_replace, "AUC"] = best_model.loc[0, "AUC"]
    ordered_results.loc[index_to_replace, "F1"] = best_model.loc[0, "F1"]

dep_var = "fuera_metas"
tgt_year = 2
base_label = 4
tgt_model = "Logistic Regression"

for tgt_samp in ["training", "validation"]:

    # Manually replace the Model type, accuracy, AUC and F1 score
    best_model = pd.read_csv(temp_path+"org_{}_years_base_{}_{}.csv".format(tgt_year, base_label, dep_var), sep = ';')
    best_model = best_model[(best_model['samp_type'] == tgt_samp) & (best_model['Model'] == tgt_model)].reset_index(drop = True)

    index_to_replace = (ordered_results['dep_var'] == dep_var) & (ordered_results['tgt_year'] == tgt_year) & (ordered_results['base_label'] == base_label) & (ordered_results['samp_type'] == tgt_samp)
    ordered_results.loc[index_to_replace, "Model"] = best_model.loc[0, "Model"]
    ordered_results.loc[index_to_replace, "Accuracy"] = best_model.loc[0, "Accuracy"]
    ordered_results.loc[index_to_replace, "AUC"] = best_model.loc[0, "AUC"]
    ordered_results.loc[index_to_replace, "F1"] = best_model.loc[0, "F1"]

### Editando las entradas como en el paper

# Change the variables' labels
ordered_results["tgt_year"].replace(to_replace = [1, 2], 
                                      value = ["1-year", 
                                               "2-year"], 
                                        inplace = True)

ordered_results["base_label"].replace(to_replace = [1, 2, 3, 4], 
                                      value = ["1 On-target", 
                                                 "2 Out-target", 
                                                 "3 Not-complications", 
                                                 "4 Yes-complications"], 
                                        inplace = True)

ordered_results["dep_var"].replace(to_replace = ["fuera_metas", "comorbilidades"], 
                                      value = ["Out-target", "Yes-complications"], 
                                        inplace = True)

ordered_results["samp_type"].replace(to_replace = ["training", "validation"], 
                                      value = ["Training", "Validation"], 
                                        inplace = True)

ordered_results['Model'].replace(to_replace = ["Logistic Regression", 
                                               "Ada Boost Classifier", 
                                               "Gradient Boosting Machine", 
                                               "Gradient Boosting Classifier"], 
                                 value = ["LR", "AdaBoost", "LGBoost", "GBC"], 
                                inplace = True)

## Export the long version of the picked models for comparisons
ordered_results.to_csv(temp_path + 'original_submission_res_long.csv', 
                       sep = ';', index = False)

# From long to wide table
reshaped_results = pd.pivot(ordered_results, 
                            index = ['tgt_year', 'base_label', 'dep_var', 'Model'], 
                            columns = "samp_type", 
                            values = [#'Model', 
                                      'Accuracy', 
                                      'F1', 
                                      'AUC']).reset_index()

# Renaming different columns
rename_dict = {"tgt_year" : "Horizon", 
               "base_label" : "Initial condition",
               "dep_var" : "Prediction"}

# Custom sorting of the table
paper_table = pd.concat([reshaped_results[(reshaped_results['dep_var'] == "Out-target") & (reshaped_results['tgt_year'] == "1-year")], 
                         reshaped_results[(reshaped_results['dep_var'] == "Yes-complications") & (reshaped_results['tgt_year'] == "1-year")],
                         reshaped_results[(reshaped_results['dep_var'] == "Out-target") & (reshaped_results['tgt_year'] == "2-year")], 
                         reshaped_results[(reshaped_results['dep_var'] == "Yes-complications") & (reshaped_results['tgt_year'] == "2-year")]]
                       ).rename(rename_dict, axis = 1)

# Final cleaning
paper_table['Initial condition'] = paper_table['Initial condition'].str.replace('\d+', '')

## Export the table in .csv format
paper_table.loc[:, [('Horizon',''),
                    ('Initial condition',''),
                    ('Prediction',''),
                    ('Model',''),
                    ("Accuracy", "Training"), 
                    ("F1", "Training"), 
                    ("AUC", "Training"),
                    ("Accuracy", "Validation"), 
                    ("F1", "Validation"), 
                    ("AUC", "Validation")]]

## Export the table in .csv format
paper_table.loc[:, [('Horizon',''),
                    ('Initial condition',''),
                    ('Prediction',''),
                    ('Model',''),
                    ("Accuracy", "Training"), 
                    ("F1", "Training"), 
                    ("AUC", "Training"),
                    ("Accuracy", "Validation"), 
                    ("F1", "Validation"), 
                    ("AUC", "Validation")]].to_csv(temp_path+'original_submission_res.csv', 
                                                   sep = ';', 
                                                   index = False)

## Generating the appendix table

# Change the variables' labels
appendix_table["tgt_year"].replace(to_replace = [1, 2], 
                                      value = ["1-year", 
                                               "2-year"], 
                                        inplace = True)

appendix_table["base_label"].replace(to_replace = [1, 2, 3, 4], 
                                      value = ["On-target", 
                                                 "Out-target", 
                                                 "Not-complicatons", 
                                                 "Yes-complications"], 
                                        inplace = True)

appendix_table["dep_var"].replace(to_replace = ["fuera_metas", "comorbilidades"], 
                                      value = ["Out-target", "Yes-complications"], 
                                        inplace = True)

appendix_table["samp_type"].replace(to_replace = ["training", "validation"], 
                                      value = ["Training", "Validation"], 
                                        inplace = True)

# Renaming different columns
rename_dict = {"tgt_year" : "Horizon", 
               "base_label" : "Initial condition",
               "dep_var" : "Prediction", 
               "samp_type" : "Sample type"}
appendix_table = appendix_table.loc[:, cols_order].rename(rename_dict, axis = 1)

appendix_table.to_csv(temp_path+"all_trained_models_metrics.csv", sep = ';', index = False)

## Generating the requested plots and tables - Reviewer 4 

### Feature importance of two models

dep_var = "fuera_metas"
tgt_year = 1
base_label = 2

print("Training horizon: {}, Initial state: {}, Outcome: {}.".format(tgt_year, base_label, dep_var))

# Generando las rutas y variables a eliminar
train_path, vali_path, vars_to_drop = parameters_gather(tgt_year = tgt_year, 
                                                        base_label = base_label,
                                                        dep_var = dep_var)

## Preparando los datos
train = data_prepare(df_path = train_path, vars_to_drop = vars_to_drop)
#train = pd.read_csv(train_path)
#train = train.loc[:, list(set(train.columns).difference(set(vars_to_drop)))]
vali = data_prepare(df_path = vali_path, vars_to_drop = vars_to_drop)

## Identificando variables categóricas y numéricas presentes
cat_vars, num_vars = cat_num_identify(ml_data = train, rename_df = rename_df)

## Renombrando las variables para el gráfico
train.rename(rename_dict, axis = 1, inplace = True)
vali.rename(rename_dict, axis = 1, inplace = True)

## Configuración del experimento
s = setup(data = train,
          target = dep_var,
          test_data = vali,
          session_id = 123,
          index = False,
          numeric_features = num_vars,
          categorical_features = cat_vars,
          remove_multicollinearity = True,
          fix_imbalance = False,
          feature_selection = True,
          #feature_selection = False,
          feature_selection_method = "sequential",
          n_features_to_select = 0.5,
          use_gpu = False, 
         # log_experiment = True
         )

# compare baseline models
best = compare_models(sort = "F1", include = ['lr'])

# Generate feature importance plot 
plot_model(best, plot = 'feature', save = True)
shutil.copyfile(".\\"+"Feature Importance.png",
                ".\\"+temp_path+"ryr_results\\baseline_replicate\\predicted_probabilities\\features_{}_years_base_{}_{}.png".format(tgt_year, base_label, dep_var))
os.remove(".\\"+"Feature Importance.png")

dep_var = "fuera_metas"
tgt_year = 1
base_label = 4

print("Training horizon: {}, Initial state: {}, Outcome: {}.".format(tgt_year, base_label, dep_var))

# Generando las rutas y variables a eliminar
train_path, vali_path, vars_to_drop = parameters_gather(tgt_year = tgt_year, 
                                                        base_label = base_label,
                                                        dep_var = dep_var)

## Preparando los datos
train = data_prepare(df_path = train_path, vars_to_drop = vars_to_drop)
#train = pd.read_csv(train_path)
#train = train.loc[:, list(set(train.columns).difference(set(vars_to_drop)))]
vali = data_prepare(df_path = vali_path, vars_to_drop = vars_to_drop)

## Identificando variables categóricas y numéricas presentes
cat_vars, num_vars = cat_num_identify(ml_data = train, rename_df = rename_df)

## Renombrando las variables para el gráfico
train.rename(rename_dict, axis = 1, inplace = True)
vali.rename(rename_dict, axis = 1, inplace = True)

## Configuración del experimento
s = setup(data = train,
          target = dep_var,
          test_data = vali,
          session_id = 123,
          index = False,
          numeric_features = num_vars,
          categorical_features = cat_vars,
          remove_multicollinearity = True,
          fix_imbalance = False,
          feature_selection = True,
          #feature_selection = False,
          feature_selection_method = "sequential",
          n_features_to_select = 0.5,
          use_gpu = False, 
         # log_experiment = True
         )

# compare baseline models
best = compare_models(sort = "F1", include = ['lr'])

# Generate feature importance plot 
plot_model(best, plot = 'feature', save = True)
shutil.copyfile(".\\"+"Feature Importance.png",
                ".\\"+temp_path+"ryr_results\\baseline_replicate\\predicted_probabilities\\features_{}_years_base_{}_{}.png".format(tgt_year, base_label, dep_var))
os.remove(".\\"+"Feature Importance.png")

### Confusion matrix of two models

## Load the probabilities matrices
ordered_results = pd.read_csv(temp_path+"ryr_results\\baseline_replicate\\predicted_probabilities\\{}.csv".format("stats_res"), 
                        sep = ';')



confusion_entries = ordered_results.loc[(ordered_results['max_kappa'] == 1), 
                    ["tgt_year", "base_label", "dep_var", "sample", "Model", "tp", "tn", "fp", "fn"]]

# Compute rates
confusion_entries['TPR'] = confusion_entries['tp']/(confusion_entries['tp']+confusion_entries['fn'])
confusion_entries['FNR'] = 1-confusion_entries['TPR']
confusion_entries['FPR'] = confusion_entries['fp']/(confusion_entries['fp']+confusion_entries['tn'])
confusion_entries['TNR'] = 1-confusion_entries['FPR']


# Change the variables' labels
confusion_entries["tgt_year"].replace(to_replace = [1, 2], 
                                      value = ["1-year", 
                                               "2-year"], 
                                        inplace = True)

confusion_entries["base_label"].replace(to_replace = [1, 2, 3, 4], 
                                      value = ["1 On-goals", 
                                                 "2 Out-goals", 
                                                 "3 Not-complications", 
                                                 "4 Yes-complications"], 
                                        inplace = True)

confusion_entries["dep_var"].replace(to_replace = ["fuera_metas", "comorbilidades"], 
                                      value = ["Out-goals", "Yes-complications"], 
                                        inplace = True)

confusion_entries["sample"].replace(to_replace = ["training", "validation"], 
                                      value = ["Training", "Validation"],
                                        inplace = True)

confusion_entries.replace(to_replace = ['LogisticRegression', 'AdaBoostClassifier', 'GradientBoostingClassifier', 'LGBMClassifier'], 
                          value = ["LR", "AdaBoost", "GBC", "LGBoost"], 
                          inplace = True)

# Renaming different columns
rename_dict = {"tgt_year" : "Horizon", 
               "base_label" : "Initial condition",
               "dep_var" : "Prediction", 
               "accuracy" : "Accuracy",
               "f1" : "F1-score", 
               "auc" : "AUC", 
               "tp" : "True Pos.", 
               "tn" : "True Neg.", 
               "fp" : "False Pos.", 
               "fn" : "False Neg."}

## Generate results for naive model
subsample = pd.pivot(confusion_entries.rename(rename_dict, axis = 1), 
                     index = ['Horizon', 'Initial condition',  'Prediction', "Model"], 
                     columns = "sample", 
                     values = ["True Pos.", 
                              "True Neg.", 
                              "False Pos.", 
                              "False Neg."]).reset_index()

# Rows sorting
subsample = pd.concat([subsample[(subsample['Prediction'] == "Out-goals") & (subsample['Horizon'] == "1-year")], 
           subsample[(subsample['Prediction'] == "Yes-complications") & (subsample['Horizon'] == "1-year")],
           subsample[(subsample['Prediction'] == "Out-goals") & (subsample['Horizon'] == "2-year")], 
           subsample[(subsample['Prediction'] == "Yes-complications") & (subsample['Horizon'] == "2-year")]])

# Final cleaning
subsample['Initial condition'] = subsample['Initial condition'].str.replace('\d+', '')
cols_order = [('Horizon',''),
              ('Initial condition',''),
              ('Prediction',''),
              ('Model','')] + [(y, x) for x in ["Training", 
                                                "Validation"] for y in ["True Pos.", 
                                                                        "True Neg.", 
                                                                        "False Pos.", 
                                                                        "False Neg."]]
subsample = subsample.loc[:, cols_order].round(4)
subsample.to_csv("ryr_results\\baseline_replicate\\predicted_probabilities\\confusion_matrices.csv", 
                 sep = ";", index = False)


print("Confusion matrix under optimal cutoff:")
subsample

## Generate results for naive model
subsample = pd.pivot(confusion_entries.rename(rename_dict, axis = 1), 
                     index = ['Horizon', 'Initial condition',  'Prediction', "Model"], 
                     columns = "sample", 
                     values = ["TPR", 
                              "FPR", 
                              "TNR", 
                              "FNR"]).reset_index()

# Rows sorting
subsample = pd.concat([subsample[(subsample['Prediction'] == "Out-goals") & (subsample['Horizon'] == "1-year")], 
           subsample[(subsample['Prediction'] == "Yes-complications") & (subsample['Horizon'] == "1-year")],
           subsample[(subsample['Prediction'] == "Out-goals") & (subsample['Horizon'] == "2-year")], 
           subsample[(subsample['Prediction'] == "Yes-complications") & (subsample['Horizon'] == "2-year")]])

# Final cleaning
subsample['Initial condition'] = subsample['Initial condition'].str.replace('\d+', '')
cols_order = [('Horizon',''),
              ('Initial condition',''),
              ('Prediction',''),
              ('Model','')] + [(y, x) for x in ["Training", 
                                                "Validation"] for y in ["TPR", 
                                                                          "FPR", 
                                                                          "TNR", 
                                                                          "FNR"]]
subsample = subsample.loc[:, cols_order].round(4)
subsample.to_csv("ryr_results\\baseline_replicate\\predicted_probabilities\\confusion_matrices_rates.csv", 
                 sep = ";", index = False)


print("Confusion matrix under optimal cutoff:")
subsample

### Barplot of F1-scores 

## Load the long panel of results
#stats_long = pd.read_csv(temp_path + 'original_submission_res_long.csv', 
#                         sep = ';')

stats_long = pd.read_csv(temp_path + "kappa_opt_long_res.csv", sep = ';')
stats_long = stats_long[stats_long['max_kappa'] == 1]

# Correct the base labels and the variables names
stats_long["base_label"].replace(to_replace = ["1 On-target", 
                                                 "2 Out-target", 
                                                 "3 Not-complications", 
                                                 "4 Yes-complications"], 
                                      value = ["On-target", 
                                                 "Out-target", 
                                                 "Not-complications", 
                                                 "Yes-complications"], 
                                        inplace = True)

stats_long.rename(columns = {"tgt_year" : "Horizon", 
                             "sample" : "Sample", 
                             "f1" : "F1 Score"}, 
                  inplace = True)

stats_long["Init. Cond.-Pred."] = stats_long['base_label']+"-"+stats_long['dep_var']

## Generate the barplot depending on the horizon 

fig = px.bar(stats_long, 
       x = "Init. Cond.-Pred.", y = "F1 Score",
       facet_row = 'Horizon', facet_col = "Sample", color = 'Init. Cond.-Pred.')

fig.update_layout(legend = dict(orientation = 'h', 
                               xanchor = 'left', 
                               x = -0.1))
fig.update_xaxes(visible=False)
pio.write_image(fig,
                temp_path+"ryr_results\\baseline_replicate\\predicted_probabilities\\"+"f1_score_per_model.jpg", 
                format="jpeg", 
                scale=1, width=1000, height=800)
